Add Ansible playbook for setting up a new user on Ubuntu
This commit introduces a new Ansible playbook that automates the process of creating a new user with administrative privileges on Ubuntu. The playbook includes tasks for user creation, adding the user to the sudo group, installing and configuring UFW firewall, generating SSH keys, and ensuring the user can SSH into the server with the generated key.